+++ pytorch_lightning/trainer/connectors/accelerator_connector.py
@@ -58,7 +58,6 @@ from pytorch_lightning.strategies import (
     DeepSpeedStrategy,
     HorovodStrategy,
     IPUStrategy,
-    ParallelStrategy,
     SingleDeviceStrategy,
     SingleTPUStrategy,
     Strategy,
@@ -783,14 +782,6 @@ class AcceleratorConnector:
     @property
     def parallel_devices(self) -> List[Union[torch.device, int]]:
         return self._parallel_devices
-
-    @property
-    def devices(self) -> int:
-        if isinstance(self.strategy, SingleDeviceStrategy):
-            return 1
-        elif isinstance(self.strategy, ParallelStrategy):
-            return len(self.strategy.parallel_devices)
-        return 0
 
     @property
     def tpu_cores(self) -> Optional[Union[List[int], int]]: